pyproject.toml
[build-system]
requires = [
  "setuptools>=64"
]
build-backend = "setuptools.build_meta"

[project]
name = "macschedule"
version = "0.1.0"
authors = [
  { name = "Jez Smith", email = "j.smith.03@cantab.net" },
]
license = { file = "LICENSE.md" }
readme = "README.md"
dependencies = [
  "click",
  "cron-converter>=1.0.0",
  "dacite>=1.8.0",
  "lxml>=5.0.0",
  "pyyaml",
]
requires-python = ">= 3.8"

[project.scripts]
msch = "macschedule.cli:cli"
